FreeCodeCamp/Python API Development_Youtube/requirements.txt
######### Main Library's ########
fastapi[all]
psycopg2-binary
sqlalchemy
passlib[bcrypt]
